Separated client library context class.
